Update to add WiFi Manager.
When there is a default config.json in LittleFS, WiFi is set in AP mode,
and the user can input their home settings. By connecting to the
"ESP8266_Setup" WiFi Access Point, a web page will prompt the user for
settings. This web page is stored within LittleFS and can be spruced up,
of course.

When all settings are verified and validated by the software, the new
settings are stored within LittleFS to replace the former default values,
and the target reboots.

If the configuration is ready at boot-up, WiFi will be set in STA mode to
process MQTT sequences.

If new settings need to be applied (new SSID, new WiFi password, etc.),
it's feasible to return to SETUP mode by pressing the C/F button 6 times
in a row and then pressing the OFF button. The system will reboot in SETUP
mode, allowing the user to change settings

Signed-off-by: Pierre-Yves MORDRET <titolini72@gmail.com>